pyproject.toml
[tool.poetry]
name = "service-monitoring"
version = "0.1.0"
description = ""
authors = ["Mateus Dotta <md.correa@outlook.com>"]
license = "MIT"

[tool.poetry.scripts]
main = "service_monitoring.main:main"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
python-dotenv = "^0.19.1"
postmarker = "^0.18.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
